+++ tests/test_operator/test_parse_accessor.py
@@ -9,6 +9,19 @@ EXAMPLES_DIR = Path(__file__).parent / "examples/accessor"
 BICEP_PARSER = BicepParser()
 
 
+def test_parse_function() -> None:
+    # given
+    sub_dir_path = EXAMPLES_DIR / "function"
+    file_path = sub_dir_path / "main.bicep"
+    expected_result = json.loads((sub_dir_path / "result.json").read_text())
+
+    # when
+    result = BICEP_PARSER.parse(file_path=file_path)
+
+    # then
+    assert_that(result).is_equal_to(expected_result)
+
+
 def test_parse_non_null() -> None:
     # given
     sub_dir_path = EXAMPLES_DIR / "non_null"